port-serial: un-schedule flash operation as soon as it's cancelled
When a flash operation is started, it is always scheduled in an idle.
If this operation is cancelled, we should right away un-schedule it,
otherwise we may end up calling flash_do() with the GTask in the
private info already gone.

See https://gitlab.freedesktop.org/mobile-broadband/ModemManager/merge_requests/178#note_330017