introspection: use correct node name for ModemManager1 object
In the D-Bus specification it is stated that the root node of
an introspection file should have an absolute path[1]. This path
should not be the root path in D-Bus as putting objects there is
considered incorrect, and also because the MM1 object does not
actually sit at the root path (instead, it uses MM_DBUS_PATH, which
is /org/freedesktop/ModemManager1). Fix this in the introspection
file.

[1] https://dbus.freedesktop.org/doc/dbus-specification.html#introspection-format

(cherry picked from commit dfc2c51b9cc7d9c780763d42c91bf696c2b8880a)