broadband-modem-qmi: reset all QMI/NET ports on initialization
If the daemon boots and it finds already existing modems in the
system (i.e. not hotplugged), we will attempt to reset all QMI/NET
ports, so that the state in which the ports are is a known one.

This logic allows us to cleanup any possible existing network link if
e.g. the daemon crashes and is restarted by the system.