bearer-qmi: create data port link if multiplex is requested/required
If the user requests or requires to use multiplexing support, try to
create a data port link and bind the WDS client(s) to it.

If the multiplex is requested but not required, we allow link creation
failures or setups where multiplexing isn't supported; we just go on
with the default connection setup without multiplexing, as long as
there is no bearer connected on the same data network interface.

If the multiplex is required and for any reason it cannot be fully
setup, we will fail the connection attempt right away.

The data port link is created in the system via the QmiDevice in the
open MMPortQmi, but it MUST be notified by the kernel (e.g. via udev
or via kernel events) before the connection attempt goes on.